fix(pokemon-fe): simplify search input state logic
